pdf-gen/requirements.txt
flask
reportlab
python-telegram-bot
requests
pypdf2